Hyper-Agents-Box/hyper-agents-box/requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
prometheus-client>=0.20.0
aiohttp
opentelemetry-sdk
opentelemetry-exporter-otlp
